+++ backend/libraries/stable_memory_map/src/keys.rs
@@ -8,6 +8,7 @@ mod community_event;
 mod expiring_event;
 mod last_updated;
 mod macros;
+mod message_activity_event;
 mod message_event_indexes;
 mod message_id;
 mod principal;
@@ -20,6 +21,7 @@ pub use chat_event::*;
 pub use community_event::*;
 pub use expiring_event::*;
 pub use last_updated::*;
+pub use message_activity_event::*;
 pub use message_event_indexes::*;
 pub use message_id::*;
 pub use principal::*;
@@ -144,6 +146,8 @@ pub enum KeyType {
     DirectChatSearchSender = 44,
     GroupChatSearchSender = 45,
     ChannelSearchSender = 46,
+    MessageActivityEvent = 47,
+    MessageActivityEventId = 48,
     #[cfg(test)]
     TestSmallEntries = 255,
 }
@@ -212,7 +216,9 @@ impl KeyType {
             | KeyType::ChannelSearchToken
             | KeyType::DirectChatSearchSender
             | KeyType::GroupChatSearchSender
-            | KeyType::ChannelSearchSender => MapClass::SmallEntries,
+            | KeyType::ChannelSearchSender
+            | KeyType::MessageActivityEvent
+            | KeyType::MessageActivityEventId => MapClass::SmallEntries,
             #[cfg(test)]
             KeyType::TestSmallEntries => MapClass::SmallEntries,
         }
@@ -289,6 +295,8 @@ impl TryFrom<u8> for KeyType {
             44 => Ok(KeyType::DirectChatSearchSender),
             45 => Ok(KeyType::GroupChatSearchSender),
             46 => Ok(KeyType::ChannelSearchSender),
+            47 => Ok(KeyType::MessageActivityEvent),
+            48 => Ok(KeyType::MessageActivityEventId),
             #[cfg(test)]
             255 => Ok(KeyType::TestSmallEntries),
             _ => Err(()),